[WebAuthn] Refactor AuthenticatorStorage
